Refactor server rate limiter
